requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
SQLAlchemy>=2.0.0
cloud-sql-python-connector[pg8000]>=1.0.0
pg8000>=1.29.0